+++ src/main/java/com/github/wintersteve25/tau/tests/TestTexture.java
@@ -10,8 +10,8 @@ import net.minecraft.client.gui.screen.inventory.InventoryScreen;
 public class TestTexture implements UIComponent {
     @Override
     public UIComponent build(Layout layout) {
-        return new Center(new Texture.Builder()
+        return new Center(new Texture.Builder(InventoryScreen.INVENTORY_LOCATION)
                 .withUvSize(new Vector2i(176, 166))
-                .build(InventoryScreen.INVENTORY_LOCATION));
+                .build());
     }
 }